added logic for calculating the total price

+++ src/main/java/com/scalux/stockManagement/dtos/ArticleDTO.java
@@ -12,7 +12,7 @@ public class ArticleDTO {
     private String designation;
     private Family family;
     private String type;
-    private Long longueur;
+    private BigDecimal longueur;
     private BigDecimal prixUnitaireHT;
     private BigDecimal prixTotalHT;
 }
